+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to the Docker Language Server will be documented in this fil
     - support lookup of `configs`, `networks`, and `secrets` referenced inside `services` object ([#91](https://github.com/docker/docker-language-server/issues/91))
   - textDocument/documentLink
     - support opening a referenced image's page as a link ([#91](https://github.com/docker/docker-language-server/issues/91))
+  - textDocument/hover
+    - extract descriptions and enum values from the Compose specification and display them as hovers ([#101](https://github.com/docker/docker-language-server/issues/101))
 
 ## [0.3.8] - 2025-04-24